+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/types/lineage/mappers/GenericLineageMapper.java
@@ -6,16 +6,16 @@ import com.linkedin.datahub.graphql.types.mappers.ModelMapper;
 import javax.annotation.Nonnull;
 import java.util.stream.Collectors;
 
-public class GenericLineageMapper implements ModelMapper<com.linkedin.common.relationships.GenericLineage, GenericLineage> {
+public class GenericLineageMapper implements ModelMapper<com.linkedin.common.lineage.GenericLineage, GenericLineage> {
 
     public static final GenericLineageMapper INSTANCE = new GenericLineageMapper();
 
-    public static GenericLineage map(@Nonnull final com.linkedin.common.relationships.GenericLineage lineage) {
+    public static GenericLineage map(@Nonnull final com.linkedin.common.lineage.GenericLineage lineage) {
         return INSTANCE.apply(lineage);
     }
 
     @Override
-    public GenericLineage apply(@Nonnull final com.linkedin.common.relationships.GenericLineage input) {
+    public GenericLineage apply(@Nonnull final com.linkedin.common.lineage.GenericLineage input) {
         final GenericLineage result = new GenericLineage();
         result.setEntities(input.getEntities().stream().map(GenericLineageRelationshipMapper::map).collect(Collectors.toList()));
         return result;
