+++ gms/api/src/main/snapshot/com.linkedin.dataset.datasets.snapshot.json
@@ -387,7 +387,7 @@
   }, "com.linkedin.common.fieldtransformer.TransformationType", "com.linkedin.common.fieldtransformer.UDFTransformer", {
     "type" : "record",
     "name" : "EntityRelationship",
-    "namespace" : "com.linkedin.common.relationships",
+    "namespace" : "com.linkedin.common.lineage",
     "doc" : "Downstream lineage information about a dataset including the source reporting the lineage",
     "fields" : [ {
       "name" : "created",
@@ -402,7 +402,7 @@
   }, {
     "type" : "record",
     "name" : "GenericLineage",
-    "namespace" : "com.linkedin.common.relationships",
+    "namespace" : "com.linkedin.common.lineage",
     "doc" : "Downstream lineage of a dataset",
     "fields" : [ {
       "name" : "entities",
@@ -1626,7 +1626,7 @@
           "name" : "genericDownstreamLineage",
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/genericDownstreamLineage",
-          "schema" : "com.linkedin.common.relationships.GenericLineage",
+          "schema" : "com.linkedin.common.lineage.GenericLineage",
           "doc" : "Rest.li entry point: /datasets/{datasetKey}/genericDownstreamLineage\n\ngenerated from: com.linkedin.metadata.resources.dataset.GenericDownstreamLineageResource",
           "simple" : {
             "supports" : [ "get" ],
