+++ metadata-models/src/main/pegasus/com/linkedin/common/lineage/EntityRelationship.pdl
@@ -1,4 +1,4 @@
-namespace com.linkedin.common.relationships
+namespace com.linkedin.common.lineage
 
 import com.linkedin.common.AuditStamp
 import com.linkedin.common.Urn
