+++ metadata-models/src/main/pegasus/com/linkedin/common/lineage/GenericLineage.pdl
@@ -1,4 +1,4 @@
-namespace com.linkedin.common.relationships
+namespace com.linkedin.common.lineage
 
 /**
  * Downstream lineage of a dataset
